requirements.txt
pyrogram
pyromod
tgcrypto
motor
dnspython
python-dotenv
telethon
asyncio
